Move string functions to string_lib.c
Ticket: None
Changelog: None
