+++ lib/leech_csv.c
@@ -8,6 +8,7 @@
 #include "files.h"
 #include "list.h"
 #include "logger.h"
+#include "string_lib.h"
 #include "utils.h"
 
 #ifdef __cplusplus
