+++ tests/unit_test.c
@@ -15,6 +15,8 @@ Suite *ListSuite(void);
 Suite *TableSuite(void);
 Suite *UtilsSuite(void);
 Suite *InstanceSuite(void);
+Suite *FilesSuite(void);
+Suite *StringLibSuite(void);
 
 int main(int argc, char *argv[]) {
   LCH_LoggerInit(LCH_LOGGER_MESSAGE_TYPE_VERBOSE_BIT |
